Implement most of the traditional where eval
I was right that it was much easier to generate the SQL clause as we
performed the yacc parse. There was a little trickery with the grammar
to avoid the 6 shift/reduce conflicts that show up in the giles query
language. It is a result of the following construction, I'm pretty sure

  whereClause  : whereTerm
               | whereClause OR whereClause
               | whereClause AND whereClause
               | LPAREN whereClause RPAREN
               ;
  (whereTerm is all terminals)

The solution is to make whereTerm *not* all terminals, and use the
LPAREN whereClause RPAREN there to define whereTerm. Then, the first
part of the OR and AND cases can be rewritten as whereTerm rather than
whereClause, which is consistent with our semantics, and doesn't cause
the shift/reduce conflicts