Implement OR using unions
This works as expected, but required a little finagling of the SQL
syntax in order to get the union to work. Namely, I think that because
you don't actually have a derived table on the other side of a UNION
(like you do with a JOIN), you can't actually apply a name to it.