Return the "valid time" to aronnax from MySQL
When a query is parsed, we decide on a canonical value for NOW that is
used throughout the query, addressing issue #2. This also gives us more
information to use to consider consistency, and might someday be
replaced with a logical timestamp. Right now, this only works when 'now'
is explicitly included or ommitted -- the queries that omit all temporal
predicates that default to "now" do not actually get a generated
timestamp, and thus default to the maximum timestamp of their returned
keys.