API reworking to handle SELECT clauses
Some of the methods have changed (Eval and EvalWhere), which allows the
downstream users (namely the interactive query line and the HTTP server)
to not have to worry about applying the SELECT clause to the array of
returned documents. EvalWhere is introduced so that the unit tests can
continue to function on evaluating whether or not the generated SQL
queries filter correctly.

Also fixed in this was a small oversight in the grammar in which a
non-temporal, non-distinct SELECT clause was not valid